refactored UserActionTest; added resend verification email test cases
