+++ src/test/java/org/onedatashare/server/system/DeleteJobTest.java
@@ -1,15 +1,39 @@
+/**
+ ##**************************************************************
+ ##
+ ## Copyright (C) 2018-2020, OneDataShare Team,
+ ## Department of Computer Science and Engineering,
+ ## University at Buffalo, Buffalo, NY, 14260.
+ ##
+ ## Licensed under the Apache License, Version 2.0 (the "License"); you
+ ## may not use this file except in compliance with the License.  You may
+ ## obtain a copy of the License at
+ ##
+ ##    http://www.apache.org/licenses/LICENSE-2.0
+ ##
+ ## Unless required by applicable law or agreed to in writing, software
+ ## distributed under the License is distributed on an "AS IS" BASIS,
+ ## WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ ## See the License for the specific language governing permissions and
+ ## limitations under the License.
+ ##
+ ##**************************************************************
+ */
+
 package org.onedatashare.server.system;
 
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.stubbing.Answer;
+import org.onedatashare.server.controller.DeleteJobController;
 import org.onedatashare.server.model.core.Job;
 import org.onedatashare.server.model.core.Role;
 import org.onedatashare.server.model.core.User;
 import org.onedatashare.server.model.request.JobRequestData;
 import org.onedatashare.server.model.useraction.UserActionResource;
 import org.onedatashare.server.repository.JobRepository;
+import org.onedatashare.server.system.base.SystemTest;
 import org.onedatashare.server.system.mockuser.WithMockCustomUser;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -26,6 +50,11 @@ import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.when;
 import static reactor.core.publisher.Mono.just;
 
+/**
+ * A system test suite that tests the deletion of {@link Job}s by their owner
+ * <br><br>
+ * Entry point for requests: {@link DeleteJobController}
+ */
 @SpringBootTest
 @RunWith(SpringRunner.class)
 @AutoConfigureMockMvc
@@ -47,6 +76,7 @@ public class DeleteJobTest extends SystemTest {
         when(userRepository.insert((User) any())).thenAnswer(addToUsers());
         when(userRepository.findById(anyString())).thenAnswer(getFromUsers());
         jobs.clear();
+        users.clear();
     }
 
     @Test
@@ -58,7 +88,7 @@ public class DeleteJobTest extends SystemTest {
         userRepository.insert(userWithJobs(TEST_USER_EMAIL, job));
 
         JobRequestData jobRequestData = jobRequestDataOf(jobId);
-        processPostWithRequestData(DELETE_JOB_URL, jobRequestData);
+        processPostWithRequestData(DELETE_JOB_URL, jobRequestData).andReturn().getAsyncResult(1000);
 
         assertTrue(jobs.get(job.getUuid()).isDeleted());
     }
@@ -77,7 +107,7 @@ public class DeleteJobTest extends SystemTest {
         userRepository.insert(userWithJobs(otherUser, otherJob));
 
         JobRequestData jobRequestData = jobRequestDataOf(jobId);
-        processPostWithRequestData(DELETE_JOB_URL, jobRequestData);
+        processPostWithRequestData(DELETE_JOB_URL, jobRequestData).andReturn().getAsyncResult(1000);
 
         assertTrue(jobs.get(job.getUuid()).isDeleted());
         assertFalse(jobs.get(otherJob.getUuid()).isDeleted());
