+++ src/test/java/org/onedatashare/server/system/LoginTest.java
@@ -1,0 +1,111 @@
+/**
+ ##**************************************************************
+ ##
+ ## Copyright (C) 2018-2020, OneDataShare Team,
+ ## Department of Computer Science and Engineering,
+ ## University at Buffalo, Buffalo, NY, 14260.
+ ##
+ ## Licensed under the Apache License, Version 2.0 (the "License"); you
+ ## may not use this file except in compliance with the License.  You may
+ ## obtain a copy of the License at
+ ##
+ ##    http://www.apache.org/licenses/LICENSE-2.0
+ ##
+ ## Unless required by applicable law or agreed to in writing, software
+ ## distributed under the License is distributed on an "AS IS" BASIS,
+ ## WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ ## See the License for the specific language governing permissions and
+ ## limitations under the License.
+ ##
+ ##**************************************************************
+ */
+
+package org.onedatashare.server.system;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.onedatashare.server.controller.LoginController;
+import org.onedatashare.server.model.core.Role;
+import org.onedatashare.server.system.base.UserActionTest;
+import org.onedatashare.server.system.mockuser.WithMockCustomUser;
+import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.junit4.SpringRunner;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+
+/**
+ * A system test suite that tests actions on user verified accounts like logging in, changing passwords.
+ * <br><br>
+ * Entry point for requests: {@link LoginController}
+ */
+@SpringBootTest
+@RunWith(SpringRunner.class)
+@AutoConfigureMockMvc
+public class LoginTest extends UserActionTest {
+
+    @Test
+    @WithMockCustomUser(username = TEST_USER_EMAIL, role = Role.USER)
+    public void givenUserVerified_WhenLoggingIn_ShouldSucceed() throws Exception {
+        registerUserAndChangePassword(TEST_USER_EMAIL, TEST_USER_PASSWORD, TEST_USER_NAME);
+
+        boolean wasSuccessful = loginUser(TEST_USER_EMAIL, TEST_USER_PASSWORD);
+
+        assertTrue(wasSuccessful);
+    }
+
+    @Test
+    @WithMockCustomUser(username = TEST_USER_EMAIL, role = Role.USER)
+    public void givenUserNotVerified_WhenLoggingIn_ShouldFail() throws Exception {
+        // does not perform verification by email
+        registerUser(TEST_USER_EMAIL, TEST_USER_NAME);
+
+        boolean loginSuccessful = loginUser(TEST_USER_EMAIL, TEST_USER_PASSWORD);
+
+        assertFalse(loginSuccessful);
+    }
+
+    @Test
+    @WithMockCustomUser(username = TEST_USER_EMAIL, role = Role.USER)
+    public void givenUserChangedPassword_WhenLoggingInWithNewPassword_ShouldSucceed() throws Exception {
+        String oldPassword = TEST_USER_PASSWORD;
+        String new_password = "new_password";
+        registerUserAndChangePassword(TEST_USER_EMAIL, oldPassword, TEST_USER_NAME);
+        loginUser(TEST_USER_EMAIL, oldPassword);
+
+        resetUserPassword(TEST_USER_EMAIL, oldPassword, new_password);
+
+        boolean loginSuccessful = loginUser(TEST_USER_EMAIL, new_password);
+        assertTrue(loginSuccessful);
+    }
+
+    @Test
+    @WithMockCustomUser(username = TEST_USER_EMAIL, role = Role.USER)
+    public void givenUserChangedPassword_WhenLoggingInWithOldPassword_ShouldFail() throws Exception {
+        String oldPassword = TEST_USER_PASSWORD;
+        String new_password = "new_password";
+        registerUserAndChangePassword(TEST_USER_EMAIL, oldPassword, TEST_USER_NAME);
+
+        resetUserPassword(TEST_USER_EMAIL, oldPassword, new_password);
+
+        boolean loginSuccessful = loginUser(TEST_USER_EMAIL, oldPassword);
+        assertFalse(loginSuccessful);
+    }
+
+    @Test
+    @WithMockCustomUser(username = TEST_USER_EMAIL, role = Role.USER)
+    public void givenIncorrectOldPassword_WhenResettingPassword_ShouldFail() throws Exception {
+        String oldPassword = TEST_USER_PASSWORD;
+        String new_password = "new_password";
+        String wrongPassword = "random_guess";
+        registerUserAndChangePassword(TEST_USER_EMAIL, oldPassword, TEST_USER_NAME);
+
+        resetUserPassword(TEST_USER_EMAIL, wrongPassword, new_password);
+
+        boolean loginSuccessful = loginUser(TEST_USER_EMAIL, wrongPassword);
+        assertFalse(loginSuccessful);
+        loginSuccessful = loginUser(TEST_USER_EMAIL, oldPassword);
+        assertTrue(loginSuccessful);
+    }
+}
